fix(wavebar): implemented bar descreasing

+++ TundraTD/Assets/Scripts/Mobs/MobModel.cs
@@ -9,15 +9,14 @@ namespace Mobs
         [SerializeField] private float maxMobHealth;
         [SerializeField] private float defaultMobDamage;
         [SerializeField] private float defaultMobSpeed;
-        [SerializeField] private int mobWaveWeight;
-        
+        [SerializeField] private float mobWaveWeight;
         private float _currentMobHealth;
         private float _currentMobDamage;
         private float _currentMobSpeed;
         private NavMeshAgent _mobNavMeshAgent;
 
         public NavMeshAgent MobNavMeshAgent => _mobNavMeshAgent;
-        public int MobWaveWeight => mobWaveWeight;
+        public float MobWaveWeight => mobWaveWeight;
         
         public float CurrentMobHealth
         {
